pyproject.toml
[tool.poetry]
name = "api-endpoint"
version = "0.1.0"
description = "Implementing various API endpoints to perform specific functions"
authors = ["BenFaruna <benjaminfaruna@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.3.3"
pytz = "^2023.3.post1"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
